[325] Move ViewExtensionDescriptionConverter to sirius-web-compatibility
Bug: https://github.com/eclipse-sirius/sirius-components/issues/325
